katycut/requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
httpx==0.27.2
openai==1.57.0
yookassa==3.1.0
python-dotenv==1.0.1
Pillow==11.0.0
pytz==2024.1
